+++ zed_components/src/zed_camera/include/sl_types.hpp
@@ -60,6 +60,7 @@
 #include <zed_msgs/msg/pos_track_status.hpp>
 #include <zed_msgs/msg/svo_status.hpp>
 #include <zed_msgs/msg/health_status_stamped.hpp>
+#include <zed_msgs/msg/heartbeat.hpp>
 #include <zed_msgs/srv/set_pose.hpp>
 #include <zed_msgs/srv/set_roi.hpp>
 #include <zed_msgs/srv/start_svo_rec.hpp>
@@ -100,6 +101,8 @@ typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::SvoStatus>>
   svoStatusPub;
 typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::HealthStatusStamped>>
   healthStatusPub;
+typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::Heartbeat>>
+  heartbeatStatusPub;
 
 typedef std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::Image>> imagePub;
 typedef std::shared_ptr<rclcpp::Publisher<stereo_msgs::msg::DisparityImage>>